requirements.txt
ffmpeg-python
faster-whisper
sentence-transformers
faiss-cpu
numpy
tqdm
pytorch 2.6+